+++ tests/test_checker/test_noqa.py
@@ -160,8 +160,9 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS439': 1,
     'WPS440': 1,
     'WPS441': 1,
-    'WPS442': 1,
+    'WPS442': 2,
     'WPS443': 1,
+    'WPS444': 1,
 
     'WPS500': 1,
     'WPS501': 1,