+++ wemake_python_styleguide/visitors/ast/blocks.py
@@ -7,12 +7,15 @@ from typing import ClassVar, DefaultDict, List, Set, Union, cast
 from typing_extensions import final
 
 from wemake_python_styleguide.compat.aliases import ForNodes, WithNodes
-from wemake_python_styleguide.logic.naming import access
 from wemake_python_styleguide.logic.naming.name_nodes import (
     flat_variable_names,
-    get_variables_from_node,
 )
 from wemake_python_styleguide.logic.nodes import get_context, get_parent
+from wemake_python_styleguide.logic.scopes import (
+    BlockScope,
+    OuterScope,
+    extract_names,
+)
 from wemake_python_styleguide.logic.walk import is_contained_by
 from wemake_python_styleguide.types import (
     AnyAssign,
@@ -21,16 +24,13 @@ from wemake_python_styleguide.types import (
     AnyImport,
     AnyNodes,
     AnyWith,
-    ContextNodes,
 )
 from wemake_python_styleguide.violations.best_practices import (
     BlockAndLocalOverlapViolation,
     ControlVarUsedAfterBlockViolation,
+    OuterScopeShadowingViolation,
 )
 from wemake_python_styleguide.visitors import base, decorators
-
-#: That's how we represent scopes that are bound to contexts.
-_ContextStore = DefaultDict[ContextNodes, Set[str]]
 
 #: That's how we represent contexts for control variables.
 _BlockVariables = DefaultDict[
@@ -85,6 +85,7 @@ class BlockVariableVisitor(base.BaseNodeVisitor):
         """
         names = {node.name} if node.name else set()
         self._scope(node, names, is_local=False)
+        self._outer_scope(node, names)
         self.generic_visit(node)
 
     def visit_any_for(self, node: AnyFor) -> None:
@@ -95,7 +96,9 @@ class BlockVariableVisitor(base.BaseNodeVisitor):
             BlockAndLocalOverlapViolation
 
         """
-        self._scope(node, _extract_names(node.target), is_local=False)
+        names = extract_names(node.target)
+        self._scope(node, names, is_local=False)
+        self._outer_scope(node, names)
         self.generic_visit(node)
 
     def visit_alias(self, node: ast.alias) -> None:
@@ -106,12 +109,10 @@ class BlockVariableVisitor(base.BaseNodeVisitor):
             BlockAndLocalOverlapViolation
 
         """
-        import_name = node.asname if node.asname else node.name
-        self._scope(
-            cast(AnyImport, get_parent(node)),
-            {import_name},
-            is_local=False,
-        )
+        parent = cast(AnyImport, get_parent(node))
+        import_name = {node.asname} if node.asname else {node.name}
+        self._scope(parent, import_name, is_local=False)
+        self._outer_scope(parent, import_name)
         self.generic_visit(node)
 
     def visit_withitem(self, node: ast.withitem) -> None:
@@ -123,11 +124,10 @@ class BlockVariableVisitor(base.BaseNodeVisitor):
 
         """
         if node.optional_vars:
-            self._scope(
-                cast(AnyWith, get_parent(node)),
-                _extract_names(node.optional_vars),
-                is_local=False,
-            )
+            parent = cast(AnyWith, get_parent(node))
+            names = extract_names(node.optional_vars)
+            self._scope(parent, names, is_local=False)
+            self._outer_scope(parent, names)
         self.generic_visit(node)
 
     # Locals:
@@ -146,6 +146,7 @@ class BlockVariableVisitor(base.BaseNodeVisitor):
             names = set(flat_variable_names([node]))
 
         self._scope(node, names, is_local=True)
+        self._outer_scope(node, names)
         self.generic_visit(node)
 
     # Utils:
@@ -157,7 +158,7 @@ class BlockVariableVisitor(base.BaseNodeVisitor):
         *,
         is_local: bool,
     ) -> None:
-        scope = _Scope(node)
+        scope = BlockScope(node)
         shadow = scope.shadowing(names, is_local=is_local)
 
         if shadow:
@@ -166,6 +167,17 @@ class BlockVariableVisitor(base.BaseNodeVisitor):
             )
 
         scope.add_to_scope(names, is_local=is_local)
+
+    def _outer_scope(self, node: ast.AST, names: Set[str]) -> None:
+        scope = OuterScope(node)
+        shadow = scope.shadowing(names)
+
+        if shadow:
+            self.add_violation(
+                OuterScopeShadowingViolation(node, text=', '.join(shadow)),
+            )
+
+        scope.add_to_scope(names)
 
 
 @final
@@ -197,7 +209,7 @@ class AfterBlockVariablesVisitor(base.BaseNodeVisitor):
 
     def visit_any_for(self, node: AnyFor) -> None:
         """Visit loops."""
-        self._add_to_scope(node, _extract_names(node.target))
+        self._add_to_scope(node, extract_names(node.target))
         self.generic_visit(node)
 
     def visit_withitem(self, node: ast.withitem) -> None:
@@ -205,7 +217,7 @@ class AfterBlockVariablesVisitor(base.BaseNodeVisitor):
         if node.optional_vars:
             self._add_to_scope(
                 cast(AnyWith, get_parent(node)),
-                _extract_names(node.optional_vars),
+                extract_names(node.optional_vars),
             )
         self.generic_visit(node)
 
@@ -239,57 +251,3 @@ class AfterBlockVariablesVisitor(base.BaseNodeVisitor):
         self.add_violation(
             ControlVarUsedAfterBlockViolation(node, text=node.id),
         )
-
-
-class _Scope(object):
-    """Represents the visibility scope of a variable."""
-
-    #: Updated when we have a new block variable.
-    _block_scopes: ClassVar[_ContextStore] = defaultdict(set)
-
-    #: Updated when we have a new local variable.
-    _local_scopes: ClassVar[_ContextStore] = defaultdict(set)
-
-    def __init__(self, node: ast.AST) -> None:
-        self._node = node
-        self._context = cast(ContextNodes, get_context(self._node))
-
-    def add_to_scope(
-        self,
-        names: Set[str],
-        is_local: bool = False,
-    ) -> None:
-        """Adds a set of names to the specified scope."""
-        scope = self._get_scope(is_local=is_local)
-        scope[self._context] = scope[self._context].union({
-            var_name  # we allow to reuse explicit `_` variable
-            for var_name in names
-            if not access.is_unused(var_name)
-        })
-
-    def shadowing(
-        self,
-        names: Set[str],
-        is_local: bool = False,
-    ) -> Set[str]:
-        """Calculates the intersection for a set of names and a context."""
-        if not names:
-            return set()
-
-        scope = self._get_scope(is_local=not is_local)
-        current_names = scope[self._context]
-
-        if not is_local:
-            # Why do we care to update the scope for block variables?
-            # Because, block variables cannot shadow
-            scope = self._get_scope(is_local=is_local)
-            current_names = current_names.union(scope[self._context])
-
-        return set(current_names).intersection(names)
-
-    def _get_scope(self, is_local: bool = False) -> _ContextStore:
-        return self._local_scopes if is_local else self._block_scopes
-
-
-def _extract_names(node: ast.AST) -> Set[str]:
-    return set(get_variables_from_node(node))