+++ CHANGELOG.md
@@ -47,6 +47,7 @@ In this release we had a little focus on:
 - Forbids to multiply lists
 - Forbids to use variable names like `__` and `_____`
 - Forbids to define unused variables explicitly
+- Forbids to shadow outer scope variables
 
 ### Bugfixes